IDEA-171404 allow UI scale exceed 2x
- UI scale can exceed 2x now
- JBUI.ScaleType.OBJ_SCALE is introduced
- JBUI.Scaler & JBUI.ScaleContext is introduced
- JBUI.JBIcon and subclasses are refactored
- CachedImageIcon is simplified
- ShadowPainter scales shadow now
- Transitioning from int/float to double where applicable
- double [x, y, width, height] is rounded this way [floor, floor, ceil, ceil]
- JBDimension is backed by double size
- JBDimension rescales itself via new methods: size()/width()/width2D()

+++ platform/util/src/com/intellij/util/SVGLoader.java
@@ -41,8 +41,8 @@ import java.util.List;
 public class SVGLoader {
   private TranscoderInput input;
   private BufferedImage img;
-  private float width;
-  private float height;
+  private double width;
+  private double height;
 
   private enum SizeAttr {
     width,
@@ -121,7 +121,7 @@ public class SVGLoader {
     return load(null, stream, scale);
   }
 
-  public static Image load(@Nullable URL url, @NotNull InputStream stream , float scale) throws IOException {
+  public static Image load(@Nullable URL url, @NotNull InputStream stream , double scale) throws IOException {
     try {
       return new SVGLoader(url, stream, scale).createImage();
     }
@@ -130,7 +130,7 @@ public class SVGLoader {
     }
   }
 
-  private SVGLoader(@Nullable URL url, InputStream stream, float scale) throws IOException {
+  private SVGLoader(@Nullable URL url, InputStream stream, double scale) throws IOException {
     Document document = null;
     String uri = null;
     try {
